+++ src/Adapter/Controller/Users/UserController.ts
@@ -1,8 +1,8 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { Request, Response } from 'express';
 import { IBaseController } from '../IBaseController';
-import CreateUserDTO from './CreateUserDTO';
 import CreateUserService from '@Application/Users/CreateUserService';
+import AppError from '@Domain/Middlewares/Errors/AppError';
 export default class UserController implements IBaseController {
   list(request: Request, response: Response): Promise<Response> {
     throw new Error('Method not implemented.');
@@ -11,13 +11,13 @@ export default class UserController implements IBaseController {
     throw new Error('Method not implemented.');
   }
   async create(request: Request, response: Response): Promise<Response> {
-    const data: CreateUserDTO = request.body;
+    const data = request.body;
 
     const createUser = new CreateUserService();
 
     await createUser.execute(data);
 
-    throw new Error('Method not implemented.');
+    throw new AppError('Method not implemented.');
   }
   update(request: Request, response: Response): Promise<Response> {
     throw new Error('Method not implemented.');
