+++ src/server.ts
@@ -1,5 +1,5 @@
+import express, { NextFunction, Request, Response } from 'express';
 import 'express-async-errors';
-import express, { Request, Response } from 'express';
 import cors from 'cors';
 import routes from '@Infrastructure/Routes';
 import AppError from '@Domain/Middlewares/Errors/AppError';
@@ -10,20 +10,22 @@ app.use(cors());
 app.use(express.json());
 app.use(routes);
 
-app.use((error: Error, request: Request, response: Response) => {
-  if (error instanceof AppError) {
-    return response.status(error.statusCode).json({
-      status: 'error',
-      message: error.message,
-    });
-  }
+app.use(
+  (error: Error, request: Request, response: Response, next: NextFunction) => {
+    if (error instanceof AppError) {
+      return response.status(error.statusCode).json({
+        status: 'error',
+        message: error.message,
+      });
+    }
 
-  return response.status(500).json({
-    status: 'error',
-    message: 'Internal Server Error',
-    innerException: error.message,
-  });
-});
+    return response.status(500).json({
+      status: 'error',
+      message: 'Internal server error',
+      innerException: error.message,
+    });
+  },
+);
 
 const PORT = process.env.PORT ? Number(process.env.PORT) : 3333;
